Arities for protocol methods now start a 0
The self parameter used to counted towards the arity of a methods,
so that one element (at 0) in the arity table for a methods was
always wasted, since all methods would have at least arity 1.

This has been fixed now, so that a method with one parameter has
arity 0 as far as the arity dispatch table is concerned.